client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "LiveKit voice client for Sir Henry"
authors = [
    {name = "Zane Reynolds", email = "zane.reynolds@gmail.com"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "livekit>=1.0.0",
    "livekit-api>=1.0.0",
    "python-dotenv>=1.0.0",
    "sounddevice>=0.4.7",
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.coverage.run]
branch = false
source = ["src"]

[tool.coverage.paths]
source = ["src"]

[tool.coverage.report]
show_missing = true
fail_under = 100
exclude_lines = [
    "# pragma: no cover",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
omit = [
]
